+++ packages/component-library-nuxt-module/CHANGELOG.md
@@ -1,3 +1,5 @@
+# [@ucla-library/component-library-nuxt-module-v1.3.34](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.3.33...@ucla-library/component-library-nuxt-module-v1.3.34) (2026-04-10)
+
 # [@ucla-library/component-library-nuxt-module-v1.3.33](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.3.32...@ucla-library/component-library-nuxt-module-v1.3.33) (2026-04-09)
 
 # [@ucla-library/component-library-nuxt-module-v1.3.32](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library/component-library-nuxt-module-v1.3.31...@ucla-library/component-library-nuxt-module-v1.3.32) (2026-04-08)